Add fallback index
@jensens, please check my comments.
I am not sure if it makes sense to name that index Language2, but I also
fear lots of issues if we replace Language with this index.
I am not confident that I found all possible edge cases, the problem
here is also, if something is a real edge case or if it can't ever happen.